Consolidate get_addresses_all().
