+++ routerd/src/zebra/address.rs
@@ -6,18 +6,31 @@
 //
 
 use std::io;
-//use std::net::{Ipv4Addr, Ipv6Addr};
+use std::str::FromStr;
+use std::net::{Ipv4Addr, Ipv6Addr};
+
 use rtable::prefix::*;
+
+pub trait AddressFamily {
+    /// Return libc Addressfamily
+    fn address_family() -> libc::c_int;
+}
+
+impl AddressFamily for Ipv4Addr {
+    fn address_family() -> libc::c_int {
+        libc::AF_INET
+    }
+}
+
+impl AddressFamily for Ipv6Addr {
+    fn address_family() -> libc::c_int {
+        libc::AF_INET6
+    }
+}
 
 /// Trait IP address handler.
 pub trait AddressHandler {
-    /// Get all IPv4 addresses from kernel.
-    fn get_ipv4_addresses_all(&self) -> Result<(), io::Error>;
-
-    /// Get all IPv6 addresses from kernel.
-    fn get_ipv6_addresses_all(&self) -> Result<(), io::Error>;
-
-
+    fn get_addresses_all<T: AddressFamily + AddressLen + FromStr>(&self) ->  Result<(), io::Error>;
 }
 
 /// Connected Address.
